+++ wifipumpkin3/core/servers/dns/pyDNSServer.py
@@ -34,7 +34,7 @@ class PyDNSServer(DNSBase):
         pass
 
     def LogOutput(self, data):
-        if self.conf.get('accesspoint', 'statusAP', format=bool):
+        if self.conf.get('accesspoint', 'status_ap', format=bool):
             self.logger.info(data)
 
     def boot(self):