+++ wifipumpkin3/core/servers/mitm/mitmmode.py
@@ -155,7 +155,7 @@ class MitmMode(Widget):
                     self.reactor.terminate()
 
     def LogOutput(self,data):
-        if self.conf.get('accesspoint', 'statusAP', format=bool):
+        if self.conf.get('accesspoint', 'status_ap', format=bool):
             self.logger.info(data)